chatbot_theme_identifier/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-multipart==0.0.9
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
pydantic-settings==2.2.1
pydantic==2.6.4
sqlalchemy==2.0.25
databases[postgresql]==0.8.0
alembic==1.13.1
celery==5.3.6
redis==5.0.1
pdfminer.six==20231228
python-multipart==0.0.9